Move calc functions to helpers.js & import
